# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Parcial 2 - Elecciones 2023'","description":"'Parcial 2 - Elecciones 2023'","contact":{"name":"Esteban Paredes","email":"estebanjparedes02@gmail.com"},"version":"'0.0.1'"},"servers":[{"url":"http://localhost:8081","description":"'Parcial 2 - Elecciones 2023'"}],"paths":{"/secciones":{"get":{"tags":["election-controller"],"operationId":"getSecciones","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"seccion_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Seccion"}}}}}}}},"/resultados":{"get":{"tags":["election-controller"],"operationId":"getResultado","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"seccion_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultadoDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/distritos":{"get":{"tags":["election-controller"],"operationId":"getDistritos","parameters":[{"name":"distrito_nombre","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Distrito"}}}}}}}},"/cargos":{"get":{"tags":["election-controller"],"operationId":"getCargosByDistrito","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DistritoCargosDTO"}}}}}}}},"components":{"schemas":{"Seccion":{"type":"object","properties":{"seccionId":{"type":"integer","format":"int64"},"seccionNombre":{"type":"string"},"distritoId":{"type":"integer","format":"int64"}}},"ResultadoDTO":{"type":"object","properties":{"distrito":{"type":"string"},"seccion":{"type":"string"},"resultados":{"type":"array","items":{"$ref":"#/components/schemas/ResultadoDetalleDTO"}}}},"ResultadoDetalleDTO":{"type":"object","properties":{"orden":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"votos":{"type":"integer","format":"int32"},"porcentaje":{"type":"number","format":"float"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"Distrito":{"type":"object","properties":{"distritoId":{"type":"integer","format":"int64"},"distritoNombre":{"type":"string"}}},"Cargo":{"type":"object","properties":{"cargoId":{"type":"integer","format":"int64"},"cargoNombre":{"type":"string"},"distritoId":{"type":"integer","format":"int64"}}},"DistritoCargosDTO":{"type":"object","properties":{"distrito":{"$ref":"#/components/schemas/Distrito"},"cargos":{"type":"array","items":{"$ref":"#/components/schemas/Cargo"}}}}}}}